--- Sources/MultipartKitTootSDK/Deprecated/MultipartError.swift
@@ -1,11 +1,0 @@
-@available(*, deprecated)
-public enum MultipartError: Error, CustomStringConvertible {
-    case invalidFormat
-    case convertibleType(Any.Type)
-    case convertiblePart(Any.Type, MultipartPart)
-    case nesting
-    case missingPart(String)
-    case missingFilename
-
-    public var description: String { "" }
-}
