--- Sources/MultipartKitTootSDK/Exports.swift
@@ -1,3 +1,0 @@
-@_exported import protocol Foundation.DataProtocol
-@_exported import struct NIO.ByteBuffer
-@_exported import struct NIOHTTP1.HTTPHeaders
