--- Sources/MultipartKitTootSDK/FormDataEncoder/FormDataEncoder.SingleValueContainer.swift
@@ -1,15 +1,0 @@
-extension FormDataEncoder.Encoder: SingleValueEncodingContainer {
-    func encodeNil() throws {
-        // skip
-    }
-
-    func encode<T: Encodable>(_ value: T) throws {
-        if let convertible = value as? MultipartPartConvertible,
-            let part = convertible.multipart
-        {
-            storage.dataContainer = SingleValueDataContainer(part: part)
-        } else {
-            try value.encode(to: self)
-        }
-    }
-}
